back-end/requirements.txt
appdirs==1.4.4
asgiref==3.2.5
astroid==2.3.3
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-cors-middleware==1.5.0
django-heroku==0.3.1
djangorestframework==3.11.0
dodgy==0.2.1
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8-naming==0.4.1
prospector==1.2.0
psycopg2==2.8.4
pycodestyle==2.4.0
pydocstyle==5.0.2
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.12
pylint-flask==0.6
pylint-plugin-utils==0.6
pytz==2019.3
PyYAML==5.3
requirements-detector==0.6
setoptconf==0.2.0
six==1.14.0
snowballstemmer==2.0.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
virtualenv==20.0.21
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.1.0
